pyproject.toml
[project]
name = "dancebreak"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "notion-client>=2.5.0",
    "pyyaml>=6.0.2",
    "notion-to-md-py>=0.1.3",
]
